+++ packages/js/CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [1.2.6] - 2020-02-06
 ### Added
-- New method `assureDeviceId()` to assure the `deviceId` is available after page refresh.
+- New method `validateDeviceId()` to assure the `deviceId` is available after page refresh.
 ### Fixed
 - Make sure `getDevices()` methods always return devices with both deviceId and label even if the browser does not have camera or microphone permissions
 ### Changed